requirements.txt
google-generativeai==0.8.3
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
mcp==1.1.1
python-multipart==0.0.9
